backend/requirements.txt
fastapi
uvicorn
python-dotenv
PyMuPDF
requests
python-docx
python-pptx
python-socketio[asgi]>=5.11.0